+++ app/src/main/java/com/example/carcamping/ui/home/BookMarkFragment.kt
@@ -1,4 +1,4 @@
-package com.example.carcamping.ui
+package com.example.carcamping.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.carcamping.R
 
-class SnapFragment : Fragment() {
+class BookMarkFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.snap_fragment, container, false)
+        return inflater.inflate(R.layout.bookmark_fragment, container, false)
     }
 
 }
