+++ app/src/main/java/com/example/carcamping/ui/home/MapFragment.kt
@@ -1,4 +1,4 @@
-package com.example.carcamping.ui
+package com.example.carcamping.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.carcamping.R
 
-class BookMarkFragment : Fragment() {
+class MapFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.bookmark_fragment, container, false)
+        return inflater.inflate(R.layout.map_fragment, container, false)
     }
 
 }
