+++ app/src/main/java/com/example/carcamping/ui/splash/SplashActivity.kt
@@ -6,7 +6,7 @@ import android.os.Handler
 import android.os.Looper
 import androidx.appcompat.app.AppCompatActivity
 import com.example.carcamping.R
-import com.example.carcamping.ui.HomeActivity
+import com.example.carcamping.ui.home.HomeActivity
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
