+++ app/src/main/res/layout/map_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.MapFragment">
+    tools:context=".ui.home.MapFragment">
 
     <TextView
         app:layout_constraintTop_toTopOf="parent"
